+++ src/publish.rs
@@ -26,6 +26,7 @@ use giganto_client::{
     },
     RawEventKind,
 };
+use log_broker::{debug, error, info, warn, LogLocation};
 use quinn::{Connection, Endpoint, RecvStream, SendStream, ServerConfig};
 use rustls::{Certificate, PrivateKey};
 use serde::{de::DeserializeOwned, Serialize};
@@ -36,7 +37,6 @@ use tokio::{
     sync::{mpsc::unbounded_channel, Notify},
     time::sleep,
 };
-use tracing::{debug, error, info, warn};
 
 const PUBLISH_VERSION_REQ: &str = ">=0.15.0,<0.17.0";
 
@@ -69,6 +69,7 @@ impl Server {
     ) {
         let endpoint = Endpoint::server(self.server_config, self.server_address).expect("endpoint");
         info!(
+            LogLocation::Both,
             "listening on {}",
             endpoint.local_addr().expect("for local addr display")
         );
@@ -90,14 +91,14 @@ impl Server {
                         )
                         .await
                         {
-                            error!("connection failed: {}", e);
+                            error!(LogLocation::Both, "connection failed: {}", e);
                         }
                     });
                 },
                 () = notify_shutdown.notified() => {
                     sleep(Duration::from_millis(SERVER_ENDPOINT_DELAY)).await;      // Wait time for channels,connection to be ready for shutdown.
                     endpoint.close(0_u32.into(), &[]);
-                    info!("Shutting down publish");
+                    info!(LogLocation::Both, "Shutting down publish");
                     break;
                 },
             }
@@ -116,11 +117,11 @@ async fn handle_connection(
 
     let (send, recv) = match server_handshake(&connection, PUBLISH_VERSION_REQ).await {
         Ok((send, recv)) => {
-            info!("Compatible version");
+            info!(LogLocation::Both, "Compatible version");
             (send, recv)
         }
         Err(e) => {
-            info!("Incompatible version");
+            info!(LogLocation::Both, "Incompatible version");
             connection.close(quinn::VarInt::from_u32(0), e.to_string().as_bytes());
             bail!("{e}")
         }
@@ -153,7 +154,7 @@ async fn handle_connection(
                 let pcap_sources = pcap_sources.clone();
                 tokio::spawn(async move {
                     if let Err(e) = handle_request(stream, db, pcap_sources).await {
-                        error!("failed: {}", e);
+                        error!(LogLocation::Both, "failed: {}", e);
                     }
                 });
             },
@@ -203,11 +204,14 @@ async fn request_stream(
                                         )
                                         .await
                                         {
-                                            error!("{}", e);
+                                            error!(LogLocation::Both, "{}", e);
                                         }
                                     }
                                     Err(_) => {
-                                        error!("Failed to deserialize hog message");
+                                        error!(
+                                            LogLocation::Both,
+                                            "Failed to deserialize hog message"
+                                        );
                                     }
                                 }
                             }
@@ -226,11 +230,14 @@ async fn request_stream(
                                         )
                                         .await
                                         {
-                                            error!("{}", e);
+                                            error!(LogLocation::Both, "{}", e);
                                         }
                                     }
                                     Err(_) => {
-                                        error!("Failed to deserialize crusher message");
+                                        error!(
+                                            LogLocation::Both,
+                                            "Failed to deserialize crusher message"
+                                        );
                                     }
                                 }
                             }
@@ -239,7 +246,7 @@ async fn request_stream(
                 }
             }
             Err(e) => {
-                error!("{}", e);
+                error!(LogLocation::Both, "{}", e);
                 break;
             }
         }
@@ -274,10 +281,13 @@ async fn process_pcap_extract(
                 // send/receive extract request from piglet
                 match pcap_extract_request(source_conn, &filter).await {
                     Ok(()) => (),
-                    Err(e) => debug!("failed to relay pcap request, {e}"),
+                    Err(e) => debug!(LogLocation::Local, "failed to relay pcap request, {e}"),
                 }
             } else {
-                error!("Failed to get {}'s connection", filter.source);
+                error!(
+                    LogLocation::Both,
+                    "Failed to get {}'s connection", filter.source
+                );
             }
         }
     });
@@ -313,10 +323,10 @@ where
                 )
                 .await
                 {
-                    error!("Failed to send network stream : {}", e);
+                    error!(LogLocation::Both, "Failed to send network stream : {}", e);
                 }
             } else {
-                error!("Failed to open conn store");
+                error!(LogLocation::Both, "Failed to open conn store");
             }
         }
         RequestStreamRecord::Dns => {
@@ -333,10 +343,10 @@ where
                 )
                 .await
                 {
-                    error!("Failed to send network stream : {}", e);
+                    error!(LogLocation::Both, "Failed to send network stream : {}", e);
                 }
             } else {
-                error!("Failed to open dns store");
+                error!(LogLocation::Both, "Failed to open dns store");
             }
         }
         RequestStreamRecord::Rdp => {
@@ -353,10 +363,10 @@ where
                 )
                 .await
                 {
-                    error!("Failed to send network stream : {}", e);
+                    error!(LogLocation::Both, "Failed to send network stream : {}", e);
                 }
             } else {
-                error!("Failed to open rdp store");
+                error!(LogLocation::Both, "Failed to open rdp store");
             }
         }
         RequestStreamRecord::Http => {
@@ -373,10 +383,10 @@ where
                 )
                 .await
                 {
-                    error!("Failed to send network stream : {}", e);
+                    error!(LogLocation::Both, "Failed to send network stream : {}", e);
                 }
             } else {
-                error!("Failed to open http store");
+                error!(LogLocation::Both, "Failed to open http store");
             }
         }
         RequestStreamRecord::Log => {
@@ -393,10 +403,10 @@ where
                 )
                 .await
                 {
-                    error!("Failed to send network stream : {}", e);
+                    error!(LogLocation::Both, "Failed to send network stream : {}", e);
                 }
             } else {
-                error!("Failed to open log store");
+                error!(LogLocation::Both, "Failed to open log store");
             }
         }
         RequestStreamRecord::Smtp => {
@@ -413,10 +423,10 @@ where
                 )
                 .await
                 {
-                    error!("Failed to send network stream : {}", e);
+                    error!(LogLocation::Both, "Failed to send network stream : {}", e);
                 }
             } else {
-                error!("Failed to open smtp store");
+                error!(LogLocation::Both, "Failed to open smtp store");
             }
         }
         RequestStreamRecord::Ntlm => {
@@ -433,10 +443,10 @@ where
                 )
                 .await
                 {
-                    error!("Failed to send network stream : {}", e);
+                    error!(LogLocation::Both, "Failed to send network stream : {}", e);
                 }
             } else {
-                error!("Failed to open ntlm store");
+                error!(LogLocation::Both, "Failed to open ntlm store");
             }
         }
         RequestStreamRecord::Kerberos => {
@@ -453,10 +463,10 @@ where
                 )
                 .await
                 {
-                    error!("Failed to send network stream : {}", e);
+                    error!(LogLocation::Both, "Failed to send network stream : {}", e);
                 }
             } else {
-                error!("Failed to open kerberos store");
+                error!(LogLocation::Both, "Failed to open kerberos store");
             }
         }
         RequestStreamRecord::Ssh => {
@@ -473,10 +483,10 @@ where
                 )
                 .await
                 {
-                    error!("Failed to send network stream : {}", e);
+                    error!(LogLocation::Both, "Failed to send network stream : {}", e);
                 }
             } else {
-                error!("Failed to open ssh store");
+                error!(LogLocation::Both, "Failed to open ssh store");
             }
         }
         RequestStreamRecord::DceRpc => {
@@ -493,10 +503,10 @@ where
                 )
                 .await
                 {
-                    error!("Failed to send network stream : {}", e);
+                    error!(LogLocation::Both, "Failed to send network stream : {}", e);
                 }
             } else {
-                error!("Failed to open dce rpc store");
+                error!(LogLocation::Both, "Failed to open dce rpc store");
             }
         }
         RequestStreamRecord::Ftp => {
@@ -513,10 +523,10 @@ where
                 )
                 .await
                 {
-                    error!("Failed to send network stream : {}", e);
+                    error!(LogLocation::Both, "Failed to send network stream : {}", e);
                 }
             } else {
-                error!("Failed to open ftp store");
+                error!(LogLocation::Both, "Failed to open ftp store");
             }
         }
         RequestStreamRecord::Mqtt => {
@@ -533,10 +543,10 @@ where
                 )
                 .await
                 {
-                    error!("Failed to send network stream : {}", e);
+                    error!(LogLocation::Both, "Failed to send network stream : {}", e);
                 }
             } else {
-                error!("Failed to open mqtt store");
+                error!(LogLocation::Both, "Failed to open mqtt store");
             }
         }
         RequestStreamRecord::Ldap => {
@@ -553,10 +563,10 @@ where
                 )
                 .await
                 {
-                    error!("Failed to send network stream : {}", e);
+                    error!(LogLocation::Both, "Failed to send network stream : {}", e);
                 }
             } else {
-                error!("Failed to open ldap store");
+                error!(LogLocation::Both, "Failed to open ldap store");
             }
         }
         RequestStreamRecord::Tls => {
@@ -573,10 +583,10 @@ where
                 )
                 .await
                 {
-                    error!("Failed to send network stream : {}", e);
+                    error!(LogLocation::Both, "Failed to send network stream : {}", e);
                 }
             } else {
-                error!("Failed to open tls store");
+                error!(LogLocation::Both, "Failed to open tls store");
             }
         }
         RequestStreamRecord::Smb => {
@@ -593,10 +603,10 @@ where
                 )
                 .await
                 {
-                    error!("Failed to send network stream : {}", e);
+                    error!(LogLocation::Both, "Failed to send network stream : {}", e);
                 }
             } else {
-                error!("Failed to open smb store");
+                error!(LogLocation::Both, "Failed to open smb store");
             }
         }
         RequestStreamRecord::Nfs => {
@@ -613,10 +623,10 @@ where
                 )
                 .await
                 {
-                    error!("Failed to send network stream : {}", e);
+                    error!(LogLocation::Both, "Failed to send network stream : {}", e);
                 }
             } else {
-                error!("Failed to open nfs store");
+                error!(LogLocation::Both, "Failed to open nfs store");
             }
         }
         RequestStreamRecord::FileCreate => {
@@ -633,10 +643,10 @@ where
                 )
                 .await
                 {
-                    error!("Failed to send sysmon stream : {}", e);
+                    error!(LogLocation::Both, "Failed to send sysmon stream : {}", e);
                 }
             } else {
-                error!("Failed to open file_create store");
+                error!(LogLocation::Both, "Failed to open file_create store");
             }
         }
         RequestStreamRecord::FileDelete => {
@@ -653,10 +663,10 @@ where
                 )
                 .await
                 {
-                    error!("Failed to send sysmon stream : {}", e);
+                    error!(LogLocation::Both, "Failed to send sysmon stream : {}", e);
                 }
             } else {
-                error!("Failed to open file_delete store");
+                error!(LogLocation::Both, "Failed to open file_delete store");
             }
         }
         RequestStreamRecord::Pcap => {}
@@ -727,7 +737,10 @@ where
             send_hog_stream_start_message(&mut sender, record_type)
                 .await
                 .map_err(|e| anyhow!("Failed to write hog start message: {}", e))?;
-            info!("start hog's publish stream : {:?}", record_type);
+            info!(
+                LogLocation::Both,
+                "start hog's publish stream : {:?}", record_type
+            );
         }
         NodeType::Crusher => {
             // crusher's policy Id always exists.
@@ -735,7 +748,10 @@ where
             send_crusher_stream_start_message(&mut sender, id)
                 .await
                 .map_err(|e| anyhow!("Failed to write crusher start message: {}", e))?;
-            info!("start crusher's publish stream : {:?}", record_type);
+            info!(
+                LogLocation::Both,
+                "start crusher's publish stream : {:?}", record_type
+            );
 
             let key_builder = StorageKey::builder()
                 .start_key(&msg.source()?)
@@ -1142,7 +1158,7 @@ async fn handle_request(
                 }
                 _ => {
                     // do nothing
-                    warn!("Not expected to reach here");
+                    warn!(LogLocation::Both, "Not expected to reach here");
                 }
             }
         }
@@ -1260,7 +1276,7 @@ async fn handle_request(
                 }
                 _ => {
                     // do nothing
-                    warn!("Not expected to reach here");
+                    warn!(LogLocation::Both, "Not expected to reach here");
                 }
             }
         }